pyproject.toml
[tool.poetry]
name = "nonebot-plugin-sell-poor"
version = "0.1.0"
description = "基于 LLM 的卖若插件"
authors = ["【夜风】NightWind <2125714976@qq.com>"]
license = "GPL3"
readme = "README.md"
packages = [{include = "nonebot_plugin_sell_poor"}]
homepage = "https://github.com/XTxiaoting14332/nonebot-plugin-sell-poor"
repository = "https://github.com/XTxiaoting14332/nonebot-plugin-sell-poor"
documentation = "https://github.com/XTxiaoting14332/nonebot-plugin-sell-poor#README.md"

[[tool.poetry.source]]
name = "tsinghua"
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple/"
default = true

[tool.poetry.dependencies]
python = ">=3.9"
nonebot2 = ">=2.3.0,<3.0.0"
httpx = ">=0.22.0"
PyJWT = ">=2.3.0"
nonebot-adapter-onebot = ">=2.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"